+++ src/Program.cs
@@ -62,6 +62,7 @@ internal class Program
             new AzureMcp.Areas.Authorization.AuthorizationSetup(),
             new AzureMcp.Areas.AzureIsv.AzureIsvSetup(),
             new AzureMcp.Areas.Cosmos.CosmosSetup(),
+            new AzureMcp.Areas.Foundry.FoundrySetup(),
             new AzureMcp.Areas.KeyVault.KeyVaultSetup(),
             new AzureMcp.Areas.Kusto.KustoSetup(),
             new AzureMcp.Areas.Monitor.MonitorSetup(),